+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bittrex parser: new data file format added.
 - Coinbase Pro parser: new "Account Statement" data file format added.
 - Coinbase parser: new "Transaction history" data file format added.
+- Coinfloor parser: new "Deposit and Withdrawal" data file format added.
 ### Changed
 - Sort wallet names in audit debug as case-insensitive.
 - Data source names in config are now case-insensitive.